refactor: split opening hours types into dedicated files
Follow the existing one-interface-per-file convention used throughout
the model directory.
