+++ packages/javascript-api/src/lib/model/opening-hours-range.ts
@@ -1,0 +1,9 @@
+import { OpeningHoursTime } from './opening-hours-time.js';
+
+/**
+ * A time range during which a location is open for business.
+ */
+export interface OpeningHoursRange {
+  opens: OpeningHoursTime;
+  closes: OpeningHoursTime;
+}
