+++ packages/javascript-api/src/lib/model/opening-hours-time.ts
@@ -1,0 +1,9 @@
+/**
+ * A time of day represented as hours and minutes.
+ */
+export interface OpeningHoursTime {
+  /** Hour of the day (0-23) */
+  hours: number;
+  /** Minute of the hour (0-59) */
+  minutes: number;
+}
